Solution to Input and Print in Hexadecimal and Octal

+++ Input and Print in Hexadecimal and Octal@@6788d5012db25efec946f941/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int num;
-    scanf("%d", &num);
+    float num;
+    scanf("%f", &num);
     printf("Hexadecimal: %x\n", num);
     printf("Octal: %o\n", num);
     return 0;
